+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [unreleased]
+### Added
+- `validate` endpoint, accepting germline or somatic json submissions and supporting validation against the official schema
 ### Changed
-- Improve README and endpoint descriptions specifying that `tsv_2_json` and `csv_2_json` endpoints support only coversion for germline submission files
+- Improve README and endpoint descriptions specifying that `tsv_2_json` and `csv_2_json` endpoints support only conversion for germline submission files
 - PR template to be more compact and user friendly
 ### Fixed
 - GitHub action to use poetry v.1.7.1